Osa 4 tehtävät tehtynä
Yhteen projektiin

+++ osa4/BlogiLista/controllers/blogs.js
@@ -1,6 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
+/* old
 const formatBlog = (blog) => {
     return {
         title: blog.title,
@@ -10,41 +13,74 @@ const formatBlog = (blog) => {
         id: blog._id
     }
 }
+*/ 
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs.map(formatBlog))
+    const blogs = await Blog
+        .find({})
+        .populate('user', {username: 1, name: 1})
+    response.json(blogs.map(Blog.format))
 
 })
 
+
 blogsRouter.post('/', async (request, response) => {
+    const body = request.body
+    
     try{
-        const body = request.body
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
 
         if(body.title === undefined || body.url === undefined){
           return response.status(400).json({error: 'title missing'})
         }
 
+        const user = await User.findById(decodedToken.id)
+
         const blog = new Blog({
           title: body.title,
           author: body.author,
           url: body.url,
-          likes: body.likes || 0
+          likes: body.likes || 0,
+          user: user._id
         })
 
-        await blog.save()
-        response.status(201).json(formatBlog(blog))
+        const savedBlog = await blog.save()
+
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
+        response.status(201).json(Blog.format(savedBlog))
     } catch (exception) {
-      console.log(exception)
-      response.status(500).json({error: 'something went wrong.!.'})
+        if(exception.name === 'JsonWebTokenError') {
+            response.status(401).send({error : exception.message})
+        } else {
+            console.log(exception)
+            response.status(500).json({error: 'something went wrong.!.'})
+        }
     }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
     try{
-        await Blog.findByIdAndRemove(request.params.id)
-        
-        response.status(204).end()
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const tokenUser = decodedToken.id
+        const blog = await Blog.findById(request.params.id)
+
+        if ( log.user.toString() === tokenUser.toString() ) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        }
+
+        return response.status(401).json({ error: 'token missing or invalid' })        
 
     } catch (exception) {
         console.log(exception)
@@ -64,7 +100,7 @@ blogsRouter.put('/:id', async (request, response) => {
 
     try{
         const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
-        response.json(formatBlog(updatedBlog))
+        response.json(Blog.format(updatedBlog))
     } catch (error) {
         console.log(error)
         response.status(400).send({error: 'malformatted id'})
